Extract reusable parts of BertModelHeadsMixin into common base class
