+++ src/transformers/adapter_heads.py
@@ -1,14 +1,26 @@
+import logging
+from abc import ABC, abstractmethod
+from typing import Union
+
 import torch
 from torch import nn
 from torch.nn import BCEWithLogitsLoss, CrossEntropyLoss, MSELoss
 
+from .adapter_composition import AdapterCompositionBlock
+from .adapter_model_mixin import ModelWithHeadsAdaptersMixin
 from .adapter_modeling import Activation_Function_Class
 from .modeling_outputs import (
     MultipleChoiceModelOutput,
     QuestionAnsweringModelOutput,
+    Seq2SeqModelOutput,
+    Seq2SeqQuestionAnsweringModelOutput,
+    Seq2SeqSequenceClassifierOutput,
     SequenceClassifierOutput,
     TokenClassifierOutput,
 )
+
+
+logger = logging.getLogger(__name__)
 
 
 # Let this class inherit from nn.Sequential to provide iterable access as before
@@ -50,9 +62,9 @@ class ClassificationHead(PredictionHead):
         }
         self.build(model)
 
-    def forward(self, outputs, attention_mask, return_dict, **kwargs):
-
-        logits = super().forward(outputs[0][:, 0])
+    def forward(self, outputs, cls_output=None, attention_mask=None, return_dict=False, **kwargs):
+        cls_output = cls_output if cls_output is not None else outputs[0][:, 0]
+        logits = super().forward(cls_output)
         loss = None
         labels = kwargs.pop("labels", None)
         if labels is not None:
@@ -65,12 +77,25 @@ class ClassificationHead(PredictionHead):
                 loss = loss_fct(logits.view(-1, self.config["num_labels"]), labels.view(-1))
 
         if return_dict:
-            return SequenceClassifierOutput(
-                loss=loss,
-                logits=logits,
-                hidden_states=outputs.hidden_states,
-                attentions=outputs.attentions,
-            )
+            if isinstance(outputs, Seq2SeqModelOutput):
+                return Seq2SeqSequenceClassifierOutput(
+                    loss=loss,
+                    logits=logits,
+                    past_key_values=outputs.past_key_values,
+                    decoder_hidden_states=outputs.decoder_hidden_states,
+                    decoder_attentions=outputs.decoder_attentions,
+                    cross_attentions=outputs.cross_attentions,
+                    encoder_last_hidden_state=outputs.encoder_last_hidden_state,
+                    encoder_hidden_states=outputs.encoder_hidden_states,
+                    encoder_attentions=outputs.encoder_attentions,
+                )
+            else:
+                return SequenceClassifierOutput(
+                    loss=loss,
+                    logits=logits,
+                    hidden_states=outputs.hidden_states,
+                    attentions=outputs.attentions,
+                )
         else:
             outputs = (logits,) + outputs[2:]
             if labels is not None:
@@ -90,8 +115,9 @@ class MultiLabelClassificationHead(PredictionHead):
         }
         self.build(model)
 
-    def forward(self, outputs, attention_mask, return_dict, **kwargs):
-        logits = super().forward(outputs[0][:, 0])
+    def forward(self, outputs, cls_output=None, attention_mask=None, return_dict=False, **kwargs):
+        cls_output = cls_output if cls_output is not None else outputs[0][:, 0]
+        logits = super().forward(cls_output)
         loss = None
         labels = kwargs.pop("labels", None)
         if labels is not None:
@@ -101,12 +127,25 @@ class MultiLabelClassificationHead(PredictionHead):
             loss = loss_fct(logits, labels)
 
         if return_dict:
-            return SequenceClassifierOutput(
-                loss=loss,
-                logits=logits,
-                hidden_states=outputs.hidden_states,
-                attentions=outputs.attentions,
-            )
+            if isinstance(outputs, Seq2SeqModelOutput):
+                return Seq2SeqSequenceClassifierOutput(
+                    loss=loss,
+                    logits=logits,
+                    past_key_values=outputs.past_key_values,
+                    decoder_hidden_states=outputs.decoder_hidden_states,
+                    decoder_attentions=outputs.decoder_attentions,
+                    cross_attentions=outputs.cross_attentions,
+                    encoder_last_hidden_state=outputs.encoder_last_hidden_state,
+                    encoder_hidden_states=outputs.encoder_hidden_states,
+                    encoder_attentions=outputs.encoder_attentions,
+                )
+            else:
+                return SequenceClassifierOutput(
+                    loss=loss,
+                    logits=logits,
+                    hidden_states=outputs.hidden_states,
+                    attentions=outputs.attentions,
+                )
         else:
             outputs = (logits,) + outputs[2:]
             if labels is not None:
@@ -126,8 +165,9 @@ class MultipleChoiceHead(PredictionHead):
         }
         self.build(model)
 
-    def forward(self, outputs, attention_mask, return_dict, **kwargs):
-        logits = super().forward(outputs[0][:, 0])
+    def forward(self, outputs, cls_output=None, attention_mask=None, return_dict=None, **kwargs):
+        cls_output = cls_output if cls_output is not None else outputs[0][:, 0]
+        logits = super().forward(cls_output)
         logits = logits.view(-1, self.config["num_choices"])
         loss = None
         labels = kwargs.pop("labels", None)
@@ -161,7 +201,7 @@ class TaggingHead(PredictionHead):
         }
         self.build(model)
 
-    def forward(self, outputs, attention_mask, return_dict, **kwargs):
+    def forward(self, outputs, cls_output=None, attention_mask=None, return_dict=False, **kwargs):
         logits = super().forward(outputs[0])
         loss = None
 
@@ -205,7 +245,7 @@ class QuestionAnsweringHead(PredictionHead):
         }
         self.build(model)
 
-    def forward(self, outputs, attention_mask, return_dict, **kwargs):
+    def forward(self, outputs, cls_output=None, attention_mask=None, return_dict=False, **kwargs):
         sequence_output = outputs[0]
         logits = super().forward(sequence_output)
         start_logits, end_logits = logits.split(1, dim=-1)
@@ -231,13 +271,27 @@ class QuestionAnsweringHead(PredictionHead):
             total_loss = (start_loss + end_loss) / 2
 
         if return_dict:
-            return QuestionAnsweringModelOutput(
-                loss=total_loss,
-                start_logits=start_logits,
-                end_logits=end_logits,
-                hidden_states=outputs.hidden_states,
-                attentions=outputs.attentions,
-            )
+            if isinstance(outputs, Seq2SeqModelOutput):
+                return Seq2SeqQuestionAnsweringModelOutput(
+                    loss=total_loss,
+                    start_logits=start_logits,
+                    end_logits=end_logits,
+                    past_key_values=outputs.past_key_values,
+                    decoder_hidden_states=outputs.decoder_hidden_states,
+                    decoder_attentions=outputs.decoder_attentions,
+                    cross_attentions=outputs.cross_attentions,
+                    encoder_last_hidden_state=outputs.encoder_last_hidden_state,
+                    encoder_hidden_states=outputs.encoder_hidden_states,
+                    encoder_attentions=outputs.encoder_attentions,
+                )
+            else:
+                return QuestionAnsweringModelOutput(
+                    loss=total_loss,
+                    start_logits=start_logits,
+                    end_logits=end_logits,
+                    hidden_states=outputs.hidden_states,
+                    attentions=outputs.attentions,
+                )
         else:
             outputs = (
                 start_logits,
@@ -246,3 +300,151 @@ class QuestionAnsweringHead(PredictionHead):
             if total_loss is not None:
                 outputs = (total_loss,) + outputs
             return outputs
+
+
+class ModelWithFlexibleHeadsAdaptersMixin(ModelWithHeadsAdaptersMixin, ABC):
+    """Adds flexible prediction heads to a model class. Implemented by the XModelWithHeads classes."""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        if not hasattr(self.config, "custom_heads"):
+            self.config.custom_heads = {}
+        self.active_head = None
+
+    def _init_head_modules(self):
+        # this dict is _only_ used for saving & reloading the configs and should not be modified otherwise
+        if not hasattr(self.config, "prediction_heads"):
+            self.config.prediction_heads = {}
+        self.heads = nn.ModuleDict(dict())
+        # add modules for heads in config
+        for head_name, config in self.config.prediction_heads.items():
+            self.add_prediction_head_from_config(head_name, config)
+
+    @abstractmethod
+    def add_prediction_head_from_config(self, head_name, config, overwrite_ok=False):
+        pass
+
+    def get_prediction_heads_config(self):
+        heads = {}
+        for head_name, head in self.heads.items():
+            heads[head_name] = head.config
+        return heads
+
+    def register_custom_head(self, identifier, head):
+        self.config.custom_heads[identifier] = head
+
+    @property
+    def active_head(self):
+        return self._active_head
+
+    @active_head.setter
+    def active_head(self, head_name):
+        self._active_head = head_name
+        if head_name is not None and head_name in self.heads:
+            self.config.label2id = self.heads[head_name].config["label2id"]
+            self.config.id2label = self.get_labels_dict(head_name)
+
+    def set_active_adapters(self, adapter_setup: Union[list, AdapterCompositionBlock]):
+        """Sets the adapter modules to be used by default in every forward pass.
+        This setting can be overriden by passing the `adapter_names` parameter in the `foward()` pass.
+        If no adapter with the given name is found, no module of the respective type will be activated.
+        In case the calling model class supports named prediction heads, this method will attempt to activate a prediction head with the name of the last adapter in the list of passed adapter names.
+
+        Args:
+            adapter_setup (list): The list of adapters to be activated by default. Can be a fusion or stacking configuration.
+        """
+        self.base_model.set_active_adapters(adapter_setup)
+        # use last adapter name as name of prediction head
+        if self.active_adapters:
+            head_name = self.active_adapters.last()
+            if head_name in self.heads:
+                self.active_head = head_name
+
+            else:
+                logger.info("No prediction head for task_name '{}' available.".format(head_name))
+
+    def add_custom_head(self, head_name, config, overwrite_ok=False):
+        if config["head_type"] in self.config.custom_heads:
+            head = self.config.custom_heads[config["head_type"]](head_name, config, self)
+            self.add_prediction_head(head, overwrite_ok)
+        else:
+            raise AttributeError(
+                "The given head as a head_type that is not registered as a custom head yet."
+                " Please register the head first."
+            )
+
+    def add_prediction_head(
+        self,
+        head: PredictionHead,
+        overwrite_ok: bool = False,
+    ):
+
+        if head.name not in self.heads or overwrite_ok:
+            self.heads[head.name] = head
+            # add reference to model config to save all head configs too
+            self.config.prediction_heads[head.name] = head.config
+
+            if "label2id" not in head.config.keys() or head.config["label2id"] is None:
+                if "num_labels" in head.config.keys():
+                    head.config["label2id"] = {"LABEL_" + str(num): num for num in range(head.config["num_labels"])}
+                if "num_choices" in head.config.keys():
+                    head.config["label2id"] = {"LABEL_" + str(num): num for num in range(head.config["num_choices"])}
+
+            logger.info(f"Adding head '{head.name}' with config {head.config}.")
+            #             self._add_prediction_head_module(head.name)
+            self.active_head = head.name
+
+        else:
+            raise ValueError(
+                f"Model already contains a head with name '{head.name}'. Use overwrite_ok=True to force overwrite."
+            )
+
+    def forward_head(
+        self, all_outputs, head_name=None, cls_output=None, attention_mask=None, return_dict=False, **kwargs
+    ):
+
+        head_name = head_name or self.active_head
+        if not head_name:
+            logger.debug("No prediction head is used.")
+            return all_outputs
+
+        if head_name not in self.heads:
+            raise ValueError("Unknown head_name '{}'".format(head_name))
+        head = self.heads[head_name]
+
+        return head(all_outputs, cls_output, attention_mask, return_dict, **kwargs)
+
+    def get_labels_dict(self, head_name=None):
+        """
+        Returns the id2label dict for the given head
+        Args:
+            head_name: (str, optional) the name of the head which labels should be returned. Default is None.
+            If the name is None the labels of the active head are returned
+
+        Returns: id2label
+
+        """
+        if head_name is None:
+            head_name = self.active_head
+        if head_name is None:
+            raise ValueError("No head name given and no active head in the model")
+        if "label2id" in self.heads[head_name].config.keys():
+            return {id_: label for label, id_ in self.heads[head_name].config["label2id"].items()}
+        else:
+            return None
+
+    def get_labels(self, head_name=None):
+        """
+        Returns the labels the given head is assigning/predicting
+        Args:
+            head_name: (str, optional) the name of the head which labels should be returned. Default is None.
+            If the name is None the labels of the active head are returned
+
+        Returns: labels
+
+        """
+        label_dict = self.get_labels_dict(head_name)
+        if label_dict is None:
+            return None
+        else:
+            return list(label_dict.values())
